requirements.txt
pyyaml
pandas
numpy
matplotlib
seaborn
sklearn
pycaret==3.2.0
lightgbm


########

pip install -r requirements.txt